+++ packages/gp-api/src/electedOffice/electedOffice.module.ts
@@ -1,12 +1,13 @@
 import { OrganizationsModule } from '@/organizations/organizations.module'
-import { Module } from '@nestjs/common'
+import { MeetingsModule } from '@/meetings/meetings.module'
+import { Module, forwardRef } from '@nestjs/common'
 import { ElectedOfficeController } from './electedOffice.controller'
 import { UseElectedOfficeGuard } from './guards/UseElectedOffice.guard'
 import { UserOrM2MGuard } from './guards/UserOrM2M.guard'
 import { ElectedOfficeService } from './services/electedOffice.service'
 
 @Module({
-  imports: [OrganizationsModule],
+  imports: [OrganizationsModule, forwardRef(() => MeetingsModule)],
   controllers: [ElectedOfficeController],
   providers: [ElectedOfficeService, UseElectedOfficeGuard, UserOrM2MGuard],
   exports: [ElectedOfficeService, UseElectedOfficeGuard],