+++ packages/gp-api/src/meetings/meetings.module.ts
@@ -1,11 +1,20 @@
-import { Module } from '@nestjs/common'
+import { Module, forwardRef } from '@nestjs/common'
+import { AgentExperimentsModule } from '@/agentExperiments/agentExperiments.module'
 import { ElectedOfficeModule } from '@/electedOffice/electedOffice.module'
+import { ElectionsModule } from '@/elections/elections.module'
+import { OrganizationsModule } from '@/organizations/organizations.module'
 import { AwsModule } from '@/vendors/aws/aws.module'
 import { MeetingsBriefingsController } from './controllers/meetingsBriefings.controller'
 import { MeetingBriefingsService } from './services/meetingBriefings.service'
 
 @Module({
-  imports: [ElectedOfficeModule, AwsModule],
+  imports: [
+    AgentExperimentsModule,
+    forwardRef(() => ElectedOfficeModule),
+    ElectionsModule,
+    OrganizationsModule,
+    AwsModule,
+  ],
   controllers: [MeetingsBriefingsController],
   providers: [MeetingBriefingsService],
   exports: [MeetingBriefingsService],